refactor: Extract MermaidDiagram component
